+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [v0.2.4] - 2026-06-16
+
+Navigation enhancements to starter kit
+
+### Added
+
+- National Nurses Week leveraging current design kit
+
+### Changed
+
+- Grouping navigation
+
 ## [v0.2.3] - 2026-06-14
 
 Code cleanup and more documentation updates